Add freeway identifier text [F4C]
We used term `motorway` instead of `freeway`. Why? `Motorway` is British
term for controled access highways. In UK and Ireland, freeways are
numbered with prefix M. This also applied to Hungary, although
`autopalya` is used to denote freeways.

In reality, all of our controled access highways are denoted as `toll
roads`, or in terms of ways, `tollways`. If we applied numbering to our
`tollways`, those should be prefixed with T. However, T prefix can also
mean trunk roads (arterial routes, possibly non freeways). To disambiguate,
`motorway` term applied to this and related signs (such as freeway route
name). The `freeway` term are used in the remaining of signs, and also
in the regulation.