Add upper black segment [P2D2]
To create the segment, first we created triangle using `Pen Tool`, which
corners are:
- intersection of #A94BB5 and  upper #ACCC4D guidelines
- upper left of blue bounding box guideline
- upper right of blue bounding box guideline
The triangle that was created then filled with black color.

Then we `Stroke and Fill to Path` the triangle by:
- converting stroke to path (`Ctrl-Alt-C`)
- break apart the stroke path from the triangle (`Ctrl-Shift-K`)
- click somewhere else, click the triangle (inside) and delete the inner
path.

Finally, we checked the segment created by zooming, to make sure that
there is no gap between the red frame circle and the segment.