Add red rectangular end marker [F4D]
To create the marker:
- Create the object which will be marked (the markee)
- (Optional) Duplicate the markee so that you can restore it later.
- Create the rectangular marker, which its width is greater than the
markee, and its height as you specified. Give the marker red fill
without stroke.
- Center-align the marker relative to the markee.
- Rotate the marker so that it touch both sides of the corner of the
markee, either clockwise or counter-clockwise. In this case, the
marker is rotated 12 degrees counter-clockwise.
- Select both the marker and the markee, then `Path -> Intersection`.
The result should be the final marker.
- Either recreate the exacly same markee, or restore the duplicate
markee created earlier to the same place as the original one.