Add outer offset (linked) triangle [P1B]
First, we resize the inner triangle to 50mm * 50mm. Then we put it on
coordinate (15mm, 15mm). We add linked offset 2mm outward the triangle
with no fill and 0.5mm black stroke. The original and linked triangle
are grouped and the resulting group are scaled to 60mm * 60mm.

Note that those methods above aren't comply to the regulation, but they
are naive ways to create the yield sign. Even MUTCD utilize them.