Add unit text [W8D]
Since the free height is defined in meters, the unit text is `m`. It
located approximately 4mm from the numeral text (distance was measured
from the bottom side of bounding box of the text). It then use Roadgeek
2005 Series 1B 80px, as in the numeral text. The text is only vertically
centered relative to the diamond frame, because it is snapped to its
location guideline.