Add blue and yellow frames [F6B]
The sign was composed by two frames: blue frame is for the name of toll
gate and its distance from the sign, and the yellow frame is for slowing
down notice.

We guessed that the size of the yellow frame is 120mm * 20mm with 1mm
black stroke and 2mm corner radii; and the size of blue frame is 120mm *
60mm with 1mm white stroke and 2mm corner radii. The gap between blue
and yellow frame is 2mm.