Add white inner stroke [F4C]
The process to create inner stroke can be found in NOTES in top root
directory.

The stroke width is half of the size in the regulation (1.5mm instead of
3mm). To make gap between the inner stroke and frame border as specified
by the regulation (2mm), we edit `inkscape:radius` value in XML editor
when the inner stroke offset was select. The value formula is (desired
gap value + half of offset stroke width). In this case it gave (2mm +
0.75mm) = 2.75mm