Add freeway name text [F4C]
The freeway name text is `CIKAMPEK-PALIMANAN` (M1), as in Wikipedia.

The drawing area for the text is 8mm from the freeway icon. The text use
Roadgeek 2005 Series E 38px. The text is centered relative to the text
drawing area, marked by default color guidelines (which have been
changed for readability) from the guideline that next to the freeway
icon.