Add red line marker [F4D]
The marker shape is rectangle with 6mm height (as mandated by the
regulation, the marker height should at least 5mm). Its width is 20mm
wider than the blue frame (the marker width is 220mm). The marker have
been centered relative to blue frame.

Note that in this process of creating `end of` sign, all other objects
beside the blue frame and red marker (that is, inner stroke, freeway
icon, and text) have been hidden in order to focus on the process of
creating marker.