Restore inner stroke [F4D]
The inner stroke is recreated the same way as creating original inner
stroke as described in NOTES file on the top-level directory.

Note that the resulting sign doesn't exactly match to the version in the
regulation. While in the regulation the red marker is cropped against
the inner stroke, in this version it is cropped against the blue frame
instead, because it will be difficult to make intersection between
inner stroke and the marker.

Also, when the frame is resized, the red marker have to be recrated again
from scratch, since it is not applicable to clip the marker to the frame
(because it will lose inner stroke).