Add context line separator in white frame [F7A-a]
First, we draw guideline for separator. This guideline is placed such a
x-coordinate that the resulting boxes later will have enough space for
exit and numeral text separately. In this case the guideline is placed
on 40mm from the left bounding box of the white frame, making 40mm wide
room for exit text and 20mm wide room for numeral text. Next, we draw a
line through intersection between the upper and lower side of white
frame with guideline, making sure that it make zero angle (vertical).
The line have stroke width matching the frame. Finally, combine the frame
and line with `Path -> Combine`.

Note that if the frame need to be widened (to accommodate numbers with
more digits [3 digits or more]), the entire frame with line separator
need to be recreated from scratch, because the process described above
permanently converts rectangle frame to path and breaks `scaling object
without scaling its corner radii and stroke width`.