Add exit name text [F7A-a]
The text is `KELUAR`. The text is placed on left-hand side of the white
frame. The text uses Roadgeek 2005 Series D 30px.

To center-align the text, first we crate 4 guidelines which snapped to
the frame's path, labeled exitFrameN, exitFrameW, exitFrameS, and
exitTextSeparator respectively. Those guidelines' color are default blue
color. Then we created no-fill-and-no-stroke centering rectangle which
stretched across the intersection of those guidelines eachother. The
text then centered against the centering rectangle. Finally, the
centering rectangle is discarded.