Recreate white frame with separator by dynamic offsetting [F7A-a]
To create frame with separators by dynamic offsetting:
- Draw frame with specified size and stroke.
- Draw one or more horizontal or vertical line separator. Make sure that
each separators run from one side of frame to the opposite side and the
stroke width and color is the same as the frame. You may need to enable
`Snap to Paths` and `Snap to Path Intersections`
- Select both frame and its separators, then `Path -> Dynamic Offset`.

To manipulate the frame (such as resizing or moving), the frame as well
as its separators need to be selected first.

The advantages over combining methods is that when resizing the frame,
the stroke is reserved (as long as scaling stroke and corner radius are
disabled).