Add text for direction to the left [D1A-b]
In the sign, the left direction text is `Simpang Dewa Ruci`. It is
placed 1.5mm from left padding area limit (signAreaW). The text use
Roadgeek 2005 Series 1B 22.5px with 3mm vertical shift (3mm lower).

Note that although the text is two lines (because of space constraints),
we choose to store it as single text object (wrapped via newline).
That is, instead of store every lines as different text objects, the text
is stored as single text object for it is the only destination that head
to the left. Any text lines that part of a destination should be
stored in single corresponding text object.