Add destinations for straight direction [D1A-b]
Destinations are:
- Denpasar
- Nusa Dua
- Sanur

Since those are three different destinations, they are stored on
different text objects. Those texts are placed with 3mm gap on each.

Because the straight direction is toward freeway, we add `Lewat Jalan
Tol` (via Freeway) text. This text's color is black and placed within
white background rectangle.

To create text with background, while align the text edge the same as
other texts:
- Create the text object. Make sure that it edges are aligned the same
as several other text objects, that is make sure that x-coordinate of
the text and other texts are same.
- Draw background rectangle which the dimension are bigger than bounding
box size of the corresponding text. Set the rectangle color different
than the text.
- Select the text and the rectangle (not vice versa), then use `Align
and Distribute` to center-align the rectangle relative to the text.
- Lower the rectangle lower than text but above the frame. The text will
appear.

All texts in this commit use Roadgeek 2005 Series 1B 22.5px. Destination
texts are left-aligned, while "via Freeway" text is center-aligned.