Resize drawing page to whole part of sign [F6B]
For viewing the file outside Inkscape and for converting to bitmap image
(like PNG and JPEG), the drawing page is resize to the size of all part
of the sign (both yellow and blue frame). This was done so that there
are no excess (unused) drawing area that would affect printing.