Intersect dashes with yellow frames [W10A]
First, we ungroup dashes and combine them. Then, we make duplicate of
the yellow frame. This was done because intersecting the frame with
dashes create new path which enclose common area of both (in this case,
the new path is dashes inside the frame). This duplicate frame then
lowered to the lowest stack in current layer. Next, select both the
dashes combination and the original frame, then `Path -> Intersection`.
The resulting intersected dashed path is then filled by red, like the
original dashes combination.