Make intersected end marker [M7A]
We have red rectangle marker and blue frame, so we can proceed directly
to intersecting them. We select both the marker and the blue circle
frame, then `Path -> Intersection`. Since the resulting intersected
marker is blue as in the original frame, we change the fill to red.
Next, we recreate 120mm * 120mm blue circle and centered it relative to
white outer frame. Finally, the object order is rearranged so that the
blue circle frame is on top of white one, and the marker is on top of
texts.

Note that due to technical reasons/restrictions, there is gap between
the marker and blue frame.