Add inner offset red triangle path [P1B]
The stroke width of red inner offset is 5mm, half of from the regulation
(10mm). The gap between outer triangle path and the offset triangle is
1.6mm.

The process of creating inner triangle path is similar to creating inner
stroke, except that when specifying offset distance. Since the stroke
width of outer triangle and the offset is different, the offset distance
formula become -(desired gap distance + (1/2 * stroke width of the outer
triangle) + (1/2 * stroke width of the offset)). In this case, the offset
distance is -(1.6mm + (1/2 * 0.75mm) + (1/2 * 5mm), which equals to
-4.475mm.