Distribute baseline of destination and distance texts [F7A-a]
In real life implementation, baseline (an imaginary line which a line of
text rests without descenders) of text objects are distributed
uniformly. To accomplish this, select all destination and distance
texts, then `Object -> Align and Distribute -> Distribute of Baseline
Texts Vertically`.

Note that if any of these texts needs to be edited, all destination and
distance texts needs to be redistributed their baselines. That is, first
vertically center-aligned those texts to the corresponding guidelines,
then use method above to redistribute baselines.