Copy F7A-a to F7A-c [F7A-c]
This sign is placed on 300 m before exit junction. Unlike F7A-a and
F7A-b, this consists of two frames, one which for exit destinations and
one for straight directions. This sign is applied to freeways which have
two or three lanes in each directions.

In this commit, F7A-a is copied to F7A-c, because the destination frames
for exit destinations have been done on F7A-a.