Add second frame [F7A-c]
This frame (straight) have same dimesion as first (exit) frame: 140mm *
90mm with 1mm white stroke and 1mm corner radii. It is placed 40mm from
the first frame. Its fill color is blue. It is placed on the same
x-coordinate as first frame (77mm without stroke, 76.5mm with stroke),
so that signAreaLimitN, signAreaLimitS, upperBoundText2,
upperBoundText3, and upperBoundDistance can be applied to the second
frame.