Left-align edges of destination texts [F7A-c]
Unwritten implementation convention states that all destination texts
have to be left-aligned their edges (bounding box) relative to the
longest (widest) text of them. In this sign, the `Soekarno-Hatta` text
is the longest of all three destination texts, so the other texts are
left-aligned to it.

To left-align edges of texts relative to the longest one:
- Select the text to be aligned. Make sure that the longest text is
select last.
- Open `Object -> Align and Distribute`.
- In "Align and Distribute" dialog, select `Last selected` on `Relative
to:` drop-down.
- Click `Align left edges`.