Add downward arrows [F7A-c]
The arrows are taken from Roadgeek 2005 Arrows 2 glyph `$`, then
converted to path and ungrouped their text group. Since this sign is for
freeways with 2 or 3 lanes on both directions, there are 2 arrows,
indicating that 2 lanes are for straight direction. Each arrow are
placed 10mm from sign area limit and snapped to bottom sign area
guideline (signAreaLimitS).

The arrow dimension is 22mm * 16.001mm.