Add freeway icon [F7A-c]
This sign have destination that lead to freeway. The destination is
`Lingkar Luar`. So we add freeway icon to the left of freeway
destination text, like in real life implementation. This is different from
Wikipedia version, which put freeway icon to the right of the text.

To create freeway icon for destination text, first we create 12mm * 16mm
blue rectangle with 0.5mm white stroke and 1mm rounded corner radii.
Then we take components of freeway icon (without its frame) from F4A,
grouped, and then resized to 11mm * 13.227mm. We center the icon to the
blue rectangle frame created earlier and group with it. The resulting
icon grouped is then placed 6mm from left bounding box of freeway name
destination text (in this case `Lingkar Luar`) and horizontal-centered
relative to the text.