Add car icon [P4G]
The car icon is taken from the regulation. It is essentially a car
divided into two sides; the front is on left-hand side, and the back is
on right-hand side, thus creating gap between those sides. The icon is
resized to 48mm * 12.532mm and centered to the circle frame, and
grouped.

Since the icon have both fill and stroke, we specify the same fill and
stroke color (black).

Note that there is gap between the stroke and the car, because the icon
is designed for the frame in the regulation, which its dimension is
differ.