Remove stroke while preserving size of car icon [P4G]
This was known as "Stroke and Fill to Path".

First, we ungrouped the car icon group. The resulting selection of paths
that make up the icon is then converted their stroke to path (`Path ->
Object to Path`). Then we break apart the stroke path and original path
(`Path -> Break Apart`). We then select individual inner paths and
deleted them.

The problem is there is a closed window path which the fill is the same
as rest of the paths when breaking apart. To workaround, after deleting
inner paths, the closed path is filled the same color as white circle,
then we selected the closed path and "back side of car" path on
right-hand side (the path surrounds the closed path), then `Path ->
Exclude`. Thus, when the inner white circle changed the fill, the closed
window path is also filled automatically.