Create unified car icon [P4G]
First, we move the front car path on left-hand side to the front of back
car path on right-hand side. The front path is snapped to the cusp node
on the bottom base (ground clearance) node of back path, then move those
paths to the same y-coordinate. We then closed the gap on the roof by
manipulating the nodes on the front path. This was done by snap the cusp
node of the roof of the front path to the back path and close bezier
nodes so that the resulting path is smooth. Finally we select both paths
and union them (`Path -> Union`).