Crop car icon to the white circle frame [F4G]
First, we duplicate the car icon so that we have two icons. Then we move
them so that one icon is on left-hand side with front part visible to
the white circle and the other icon is on right-hand side which the rear
part is on the circle. Those icons are horizontally-centered to the
white frame. Next, we select the left car path and the white circle, and
`Path -> Intersection`. Note that this process will created cropped car
path with white circle disappeared, so we need to create the circle
again with the same dimension as the original one and centered to the
red circle. This white circle is used to intersect with right car path.

After two cropped car path are created, we draw again the white circle,
with the same dimension as the white circle before intersection process.

Note that when drawing white circle after intersecting, it is placed
above the red circle frame in order for the car paths to be visible.