Add circle frames [M1A, M1B, M1C & M1D]
We use stacked technique. That is, the frame is created by drawing outer
and inner frame.

The outer circle frame size is 64mm * 64mm with 0.1mm black stroke and
white color. The inner circle frame is 60mm * 60mm size with blue fill
and without stroke and centered to the white outer frame. This was done
to create 2mm gap between outer and inner frame.