Add yellow diamond frame [W1A & W1B]
The diamond dimension is 60mm * 60mm with 2.5mm black stroke and 3.7mm
rounded corner radii.

To create diamond, first we draw rectangle with specified rounded
corner radii. Then we rotate the rectangle by 45 degrees either
clockwise or counter-clockwise. The resulting diamond is then resized
using Pick Tool to the specified width and height. Note that we need to
iteratively resize because the resulting sizes interpolate from the
desired size.