Add meter numeral text [F5B]
First, we edit meterNumeralHeight guideline name to
meterNumeralPaddingHeight, as this guideline is used as padding for
meter numeral text.

The placeholder text for meter numeral is `999`. It use Roadgeek 2005
Series D font, white color. The font size spanning from
meterNumeralPaddingHeight to signPaddingS guidelines is 40.845px, thus
we round it to next even number (42px). The meter numeral text is
centered to green part of sign frame.