Distribute baseline of texts [F6C]
The texts consist of two columns: the left-hand side is for exit literal
and destination text; the right-hand side is for distance texts. For
each column, we distribute baseline of texts on that column.

However, because baseline of texts are nearly equal before applying
baseline distribution, not all texts are shifted horizontally; if
shifted, only slightly. Note that if any of texts (except text literals)
are edited, all texts have to be horizontally-centered to their bounding
box guidelines first, then re-apply baseline distribution.