Add hospital path objects [D6B1]
The objects consisted of red cross (plus sign) and white bed.

The red cross size is 10mm * 9.962mm and placed 5mm from upper and right
border of frame (that is, placed on top-right side).

The white bed size is 40mm * 17.480mm and placed 5mm from bottom border
of frame, and vertically-centered by placing it on 5mm from left border
of frame.

Both objects are copied from the regulation.