Add barrier text [P1C]
The text define which facility the mandated activity must be done on. In
this sign, `GERBANG TOL` (toll gate) is used. This means that traffic
must pay toll to resume the flow on the tolled freeways.

The text use Roadgeek 2005 Series 1B 22px with 6mm tracking space
between letters and centered to the frame. The text is placed 2mm from
the rectangle.